POST /api/v1/hospitals/{id}/reviews GET /api/v1/hospitals/reviews/{id} 1. [POST] 리뷰 등록기능 추가 2. [GET] 리뷰 조회기능 추가

+++ src/main/java/com/jpa/exercise/domain/dto/HospitalItemResponse.java
@@ -1,6 +1,5 @@
 package com.jpa.exercise.domain.dto;
 
-
 import com.jpa.exercise.domain.Review;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -15,8 +14,9 @@ import java.util.List;
 @Builder
 public class HospitalItemResponse {  // DTO
 
-    private Integer id;
+    private Long id;
     private String roadNameAddress;
     private String hospitalName;
     private List<Review> reviews;
+
 }
