+++ src/main/java/com/jpa/exercise/domain/dto/ReviewCreateRequest.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @Getter
 public class ReviewCreateRequest {
 
-    private Integer hospitalId;
+    private Long hospitalId;
     private String title;
     private String content;
     private String patientName;
